requirements.txt
lxml
requests
bs4
mysql-connector-python